+++ auth-service/src/main/java/com/sparta/blackyolk/auth_service/AuthServiceApplication.java
@@ -1,0 +1,15 @@
+package com.sparta.blackyolk.auth_service;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.PropertySource;
+
+@SpringBootApplication
+@PropertySource("classpath:env.properties")
+public class AuthServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(AuthServiceApplication.class, args);
+	}
+
+}
