+++ auth-service/src/main/java/com/sparta/blackyolk/auth_service/user/dto/DeleteResponseDto.java
@@ -1,0 +1,16 @@
+package com.sparta.blackyolk.auth_service.user.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.time.LocalDateTime;
+
+@Getter
+@Setter
+@Builder
+public class DeleteResponseDto {
+    private Long userId;
+    private Boolean isDeleted;
+    private LocalDateTime deletedAt;
+}
