+++ auth-service/src/main/java/com/sparta/blackyolk/auth_service/user/dto/LoginRequestDto.java
@@ -1,0 +1,11 @@
+package com.sparta.blackyolk.auth_service.user.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Setter
+@Getter
+public class LoginRequestDto {
+    private String username;
+    private String password;
+}
