+++ auth-service/src/main/java/com/sparta/blackyolk/auth_service/user/repository/UserRepositoryCustom.java
@@ -1,0 +1,9 @@
+package com.sparta.blackyolk.auth_service.user.repository;
+
+import com.sparta.blackyolk.auth_service.user.entity.User;
+
+import java.util.Optional;
+
+public interface UserRepositoryCustom {
+    Optional<User> findUserByDynamicCondition(Long userId, String username, String email, String slackId);
+}
