--- auth-service/src/main/java/com/sparta/msa_exam/auth/auth_service/AuthServiceApplication.java
@@ -1,15 +1,0 @@
-package com.sparta.msa_exam.auth.auth_service;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.PropertySource;
-
-@SpringBootApplication
-@PropertySource("classpath:env.properties")
-public class AuthServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(AuthServiceApplication.class, args);
-	}
-
-}
