--- auth-service/src/main/java/com/sparta/msa_exam/auth/auth_service/config/PasswordConfig.java
@@ -1,15 +1,0 @@
-package com.sparta.msa_exam.auth.auth_service.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-@Configuration
-public class PasswordConfig {
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-}
