--- auth-service/src/main/java/com/sparta/msa_exam/auth/auth_service/security/UserDetailsServiceImpl.java
@@ -1,26 +1,0 @@
-package com.sparta.msa_exam.auth.auth_service.security;
-
-import com.sparta.msa_exam.auth.auth_service.user.entity.User;
-import com.sparta.msa_exam.auth.auth_service.user.repository.UserRepository;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserDetailsServiceImpl implements UserDetailsService {
-
-    private final UserRepository userRepository;
-
-    public UserDetailsServiceImpl(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        User user = userRepository.findByUsername(username)
-                .orElseThrow(() -> new UsernameNotFoundException("Not Found " + username));
-
-        return new UserDetailsImpl(user);
-    }
-}
