--- auth-service/src/main/java/com/sparta/msa_exam/auth/auth_service/user/dto/LoginRequestDto.java
@@ -1,11 +1,0 @@
-package com.sparta.msa_exam.auth.auth_service.user.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Setter
-@Getter
-public class LoginRequestDto {
-    private String username;
-    private String password;
-}
