--- auth-service/src/main/java/com/sparta/msa_exam/auth/auth_service/user/dto/LoginResponseDto.java
@@ -1,12 +1,0 @@
-package com.sparta.msa_exam.auth.auth_service.user.dto;
-
-import com.sparta.msa_exam.auth.auth_service.user.entity.UserRoleEnum;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class LoginResponseDto {
-    private Long userId; // 사용자 ID
-    private UserRoleEnum role; // 사용자 역할 (ENUM 타입 유지)
-}
