+++ gateway/build.gradle
@@ -34,6 +34,7 @@ dependencies {
 	testImplementation 'org.springframework.boot:spring-boot-starter-test' // 테스트용 의존성
 	testImplementation 'io.projectreactor:reactor-test' // 리액티브 테스트
 	testRuntimeOnly 'org.junit.platform:junit-platform-launcher' // JUnit 런처
+	implementation 'org.springframework.boot:spring-boot-starter-web'
 }
 
 dependencyManagement {
