+++ gateway/src/main/java/com/sparta/blackyolk/gateway/config/SecurityConfig.java
@@ -1,0 +1,38 @@
+package com.sparta.blackyolk.gateway.config;
+
+import com.sparta.blackyolk.gateway.LocalJwtAuthenticationFilter;
+import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+
+@Configuration
+@EnableWebSecurity // Spring Security 지원을 가능하게 함
+@EnableMethodSecurity(securedEnabled = true)
+public class SecurityConfig {
+
+    @Bean
+    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+        // CSRF 비활성화
+        http.csrf(csrf -> csrf.disable());
+
+        // JWT를 사용하는 경우 세션 비활성화
+        http.sessionManagement(sessionManagement ->
+                sessionManagement.sessionCreationPolicy(SessionCreationPolicy.STATELESS)
+        );
+
+        // 경로별 권한 설정
+        http.authorizeHttpRequests(authorize ->
+                authorize
+                        .requestMatchers("/api/users/signup", "/api/users/login").permitAll() // 회원가입, 로그인 경로는 인증 제외
+                        .anyRequest().authenticated() // 나머지 요청은 인증 필요
+        );
+
+        return http.build();
+    }
+}
